src/requirements.txt
PySide6>=6.5
mysql-connector-python>=8.0
reportlab>=4.0
matplotlib>=3.5